pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Cohere Application Toolkit"
authors = ["Cohere"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.109.2"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
sqlalchemy = "^2.0.26"
pydantic = "^2.6.4"
langchain = "0.1.16"
wikipedia = "^1.4.0"
python-dotenv = "^1.0.1"
pytest-dotenv = "^0.5.2"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
sse-starlette = "^2.0.0"
boto3 = "^1.0.0"
httpx = "^0.27.0"
chromadb = "^0.4.16"
cohere = "^5.5.7"
llama-index = "^0.10.11"
inquirer = "^3.2.4"
langchain-community = "^0.0.32"
langchain-cohere = "^0.1.1"
langchain-experimental = "^0.0.56"
py-expression-eval = "^0.3.14"
tavily-python = "^0.3.3"
arxiv = "^2.1.0"
xmltodict = "^0.13.0"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"
bcrypt = "^4.1.2"
pypdf = "^4.2.0"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.1"
transformers = "^4.41.2"
google_auth_oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.133.0"
openpyxl = "^3.1.5"
redis = {extras = ["hiredis"], version = "^5.0.7"}
python-docx = "^1.1.2"
python-calamine = "^0.2.3"

# Compass dependencies - To be removed once Compass is OSS
structlog = "^24.4.0"
pyyaml = "^6.0.1"
[tool.poetry.group.compass]
optional = false

[tool.poetry.group.compass.dependencies]
fsspec = "2024.2.0"
joblib = "*"
pydantic = ">=2.6.3"
python = ">=3.9,<3.12"
requests = ">=2.25.0,<3.0.0"
tenacity = "8.2.3"
tqdm = ">=4.42.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
pytest-env = "^1.1.3"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
pre-commit = "^2.20.0"
ruff = "^0.0.94"
isort = "^5.12.0"
black = "24.3.0"
autoflake="^2.3.1"
pytest-asyncio = "^0.23.7"

[tool.poetry.group.setup]
optional = true

[tool.poetry.group.setup.dependencies]
python = "~3.11"
inquirer = "3.2.4"
python-dotenv = "1.0.1"
cohere = "^5.0.2"
sqlalchemy = "^2.0.26"
fastapi = "^0.109.2"
psycopg2 = "^2.9.9"
boto3 = "^1.0.0"
py-expression-eval = "^0.3.14"
langchain = "0.1.16"
langchain-cohere = "^0.1.1"
tavily-python = "^0.3.3"

[tool.poetry.group.community]
optional = true

[tool.poetry.group.community.dependencies]
wolframalpha = "^5.0.0"
torch = "^2.3.0"
langchain-mongodb = "0.1.6"

[tool.poetry.group.local-model]
optional = true

[tool.poetry.group.local-model.dependencies]
llama-cpp-python = "^0.2.67"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pyright]
include = ["src/backend/services/metrics.py"]
defineConstant = { DEBUG = true }
reportMissingImports = true
reportMissingTypeStubs = false
